+++ src/main/java/sillenceSoft/schedulleCall/Dto/StatusDto.java
@@ -8,8 +8,8 @@ import java.time.LocalDateTime;
 @Data
 @Builder
 public class StatusDto {
-    private Integer userNo;
-    private Integer statusNo;
+    private Long userNo;
+    private Long statusNo;
     private String status;
     private LocalDateTime modDt;
     private boolean isFromSchedule;
