+++ src/main/java/sillenceSoft/schedulleCall/Repository/AccessRepository.java
@@ -12,8 +12,8 @@ import java.util.List;
 
 @Mapper
 public interface AccessRepository  {
-    List<String> getAccessList (Integer userNo);
-    void saveAccess (@Param(value = "userNo") Integer userNo, @Param(value = "accessUserPhone") String accessUserPhone);
-    void deleteAccess (@Param(value = "userNo")Integer userNo, @Param(value = "accessUserPhone")String accessUserPhone);
-    Integer checkAccessOrNot (@Param(value="userNo") Integer userNo, @Param(value="accessUserPhone") String accessUserPhone );
+    List<String> getAccessList (Long userNo);
+    void saveAccess (@Param(value = "userNo") Long userNo, @Param(value = "accessUserPhone") String accessUserPhone);
+    void deleteAccess (@Param(value = "userNo")Long userNo, @Param(value = "accessUserPhone")String accessUserPhone);
+    Long checkAccessOrNot (@Param(value="userNo") Long userNo, @Param(value="accessUserPhone") String accessUserPhone );
 }
